requirements.txt
flask
ultralytics
opencv-python
numpy
pillow
torch
torchvision
torchaudio
Flask-Cors